+++ src/main/java/harmonised/pmmo/gui/CreditorScreen.java
@@ -72,7 +72,7 @@ public class CreditorScreen extends Screen
     }
 
     @Override
-    public void render( int mouseX, int mouseY, float partialTicks)
+    public void render( int mouseX, int mouseY, double partialTicks)
     {
         renderBackground( 1 );
         super.render( mouseX, mouseY, partialTicks );
