+++ src/main/java/harmonised/pmmo/gui/ListScreen.java
@@ -1010,7 +1010,7 @@ public class ListScreen extends Screen
     }
 
     @Override
-    public void render(int mouseX, int mouseY, float partialTicks)
+    public void render(int mouseX, int mouseY, double partialTicks)
     {
         renderBackground( 1 );
