+++ src/main/java/harmonised/pmmo/gui/PrefsScrollPanel.java
@@ -74,7 +74,7 @@ public class PrefsScrollPanel extends ScrollPanel
                 }
                 else
                 {
-                    if( (float) prefEntry.defaultVal == (float) slider.getValue() )
+                    if( (double) prefEntry.defaultVal == (double) slider.getValue() )
                         fillGradient(this.left + 4, prefEntry.y - 11, this.right - 2, prefEntry.y + slider.getHeight() + 2, 0xaa444444, 0xaa222222);
                     else
                         fillGradient(this.left + 4, prefEntry.y - 11, this.right - 2, prefEntry.y + slider.getHeight() + 2, 0x22444444, 0x33222222);
@@ -121,7 +121,7 @@ public class PrefsScrollPanel extends ScrollPanel
     }
 
     @Override
-    public void render(int mouseX, int mouseY, float partialTicks)
+    public void render(int mouseX, int mouseY, double partialTicks)
     {
 //        this.drawBackground();
 
@@ -149,7 +149,7 @@ public class PrefsScrollPanel extends ScrollPanel
 //            RenderSystem.disableFog();
 //            this.client.getTextureManager().bindTexture(AbstractGui.BACKGROUND_LOCATION);
 //            RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
-//            final float texScale = 32.0F;
+//            final double texScale = 32.0F;
 //            worldr.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX_COLOR);
 //            worldr.pos(this.left,  this.bottom, 0.0D).tex(this.left  / texScale, (this.bottom + (int)this.scrollDistance) / texScale).color(0x20, 0x20, 0x20, 0xFF).endVertex();
 //            worldr.pos(this.right, this.bottom, 0.0D).tex(this.right / texScale, (this.bottom + (int)this.scrollDistance) / texScale).color(0x20, 0x20, 0x20, 0xFF).endVertex();
