+++ src/main/java/harmonised/pmmo/gui/StatsScreen.java
@@ -64,7 +64,7 @@ public class StatsScreen extends Screen
     }
 
     @Override
-    public void render(int mouseX, int mouseY, float partialTicks)
+    public void render(int mouseX, int mouseY, double partialTicks)
     {
         renderBackground( 1 );
         super.render(mouseX, mouseY, partialTicks);
